chatterpy_requirements.txt
langchain-community==0.2.5
torch==2.2.2
ibm_watson_machine_learning==1.0.359
ibm_watsonx_ai==1.0.8
langchain-ibm==0.1.7
pypdf==4.3.1
Wikipedia-API==0.6.0
langchain-ollama==0.1.1
qdrant-client==1.11.0
langchain-qdrant==0.1.3
streamlit==1.37.1